Remove CommitFinder.setMatcher
This is required to upgrade to JGit 2.0 which changed
the order by which commit filter and tree filter are
invoked when both are being used in a RevWalk.

The commit filter is no longer called since the tree
filter is earlier in the chain and therefore having
two commit filters that operate before and after the
tree filter is no longer possible.

A CommitParentTreeFilter can now be used to create a
commit filter that includes commits where a tree filter
includes at least one path from a tree walk that includes
the current commit's tree and all the parent trees.